+++ Changelog.md
@@ -5,6 +5,7 @@
 ### Features and Improvements
 
 - Add Automations page - [#695](https://github.com/PrefectHQ/ui/pull/695)
+- Add Service Accounts and API keys to the UI - [#678](https://github.com/PrefectHQ/ui/pull/678)
 - Add generic markdown messages for action notifications - [#674](https://github.com/PrefectHQ/ui/pull/674)
 
 ### Bugfixes